+++ h5features/writer.py
@@ -41,13 +41,23 @@ class Writer(object):
     :param char mode: Optional. The mode for overwriting an existing
         file, 'a' to append data to the file, 'w' to overwrite it
 
+    :param str_or_int_or_none compression: Optional. Whether to
+        compress the data or not. If None, do not compress. If a
+        string, must be 'gzip' or 'lzf'. If an integer, uses 'gzip'
+        and indicates the compression level, must be from 0 to 9 (if
+        'gzip' specified, the compression level is is 4). 'gzip' has
+        good compression, moderate speed, 'lzf' has low to moderate
+        compression, very fast speed. If appending data to an existing
+        uncompressed group, data will not be compressed.
+
     :raise IOError: if the file exists but is not HDF5, if the file
         can be opened, if the mode is not 'a' or 'w', if the chunk
-        size is below 8 Ko or if the requested version is not
-        supported.
+        size is below 8 Ko, if the requested version is not
+        supported or if the specified compression is not valid.
 
     """
-    def __init__(self, filename, chunk_size=0.1, version='1.1', mode='a'):
+    def __init__(self, filename, chunk_size=0.1, version='1.1',
+                 mode='a', compression=None):
         # check version
         if not is_supported_version(version):
             raise IOError('version {} is not supported'.format(version))
@@ -62,7 +72,7 @@ class Writer(object):
         if not isinstance(chunk_size, numbers.Number):
             raise IOError(
                 'chunk size must be a number, it is {}'.format(
-                    chunk_size.__class__))
+                    chunk_size.__class__.__name__))
         if chunk_size < 0.008:
             raise IOError('chunk size is below 8 Ko')
         self.chunk_size = chunk_size
@@ -71,6 +81,27 @@ class Writer(object):
         if mode not in ('w', 'a'):
             raise IOError(
                 "mode for writing must be 'a' or 'w', it is '{}'".format(mode))
+
+        # check compression
+        if compression is None:
+            self.compression = None
+            self.compression_opts = None
+        elif isinstance(compression, str):
+            if compression not in ('gzip', 'lzf'):
+                raise IOError(
+                    "compression must be 'gzip' or 'lzf', it is '{}'"
+                    .format(compression))
+            self.compression = compression
+            self.compression_opts = None
+        elif isinstance(compression, int):
+            if not 0 <= compression <= 9:
+                raise IOError(
+                    'compression level must be in [0, 9], it is {}'
+                    .format(compression))
+            self.compression = 'gzip'
+            self.compression_opts = compression
+        else:
+            raise IOError("compression must be None, 'gzip', 'lzf' or an int")
 
         try:
             self.h5file = h5py.File(self.filename, mode=mode)
@@ -126,5 +157,6 @@ class Writer(object):
             del self.h5file[groupname]
         group = self.h5file.create_group(groupname)
         group.attrs['version'] = self.version
-        data.init_group(group, self.chunk_size)
+        data.init_group(
+            group, self.chunk_size, self.compression, self.compression_opts)
         return group